requirements.txt
Pillow==7.2.0
streamlit==0.64.0
